chore: rename views folder to pages
